implement new method of starting winbar, add show and hide subcommands (#1)
This commit fixes winbar not continuing to run after being started with winbarc and closing the terminal. An introduction of a new way to start winbar was implemented to remedy this. As a result of this, an issue arose in development where the status bar wouldn't properly be shown if winbarc was run using cargo run in debug mode. For now, release mode will need to be specified if choosing to run winbarc in this way.

A fix was also made to the status bar being invisible when winbar was started in some cases. See #1 for details.

Two new winbarc subcommands were also implemented for hiding and showing the status bar.