replace option of color with colorconfig variant, fully implement inline styles
This commit makes a change to optional colors such that instead of specifying an Option<ColorConfig>, one would use ColorConfig::Default. This comes as a result of implementing inline styles, where a ColorConfig can be deserialized either from a string or map. In particular, it is trivial to parse a ColorConfig, however, Option<ColorConfig> is not as trivial, and thus, the easiest way forward was to add a 'None' type to ColorConfig.

This change to ColorConfig does not bleed into the internal Color type (panics if trying to convert ColorConfig::Default into Color). If necessary, reverting this particular change can be looked at in the future.