feat: styling (#3)
This commit introduces initial work on overhauling styling in winbar. See #3 for full changes.

Commits:
* add style options, migrate datetime component to new style system, add new rect type

* fix wrong component location calculations, custom font and font size not being used in datetime component

* migrate static text component to new style system

* fix clippy warnings

* add status bar background color config option

* centralize style setting

* rename default bg and fg color in config to be more descriptive, regenerate example config

* start hex parsing

* fully implement parsing hex colors and numbers

* separate internal types from config types

Previously, internal and config types were shared. They are now separate to allow more flexibility on what we can accept on the config side.

* start inline colour parsing

* replace option of color with colorconfig variant, fully implement inline styles

This commit makes a change to optional colors such that instead of specifying an Option<ColorConfig>, one would use ColorConfig::Default. This comes as a result of implementing inline styles, where a ColorConfig can be deserialized either from a string or map. In particular, it is trivial to parse a ColorConfig, however, Option<ColorConfig> is not as trivial, and thus, the easiest way forward was to add a 'None' type to ColorConfig.

This change to ColorConfig does not bleed into the internal Color type (panics if trying to convert ColorConfig::Default into Color). If necessary, reverting this particular change can be looked at in the future.

* cleanup